[refactor] 패키지 구조 리팩토링 및 공통 설정 개선 - Lol Esports API WebClient에 default header(x-api-key) 등록   → API 요청마다 개별 header 추가 필요 제거 - 전체 패키지 구조 도메인 기반으로 리팩토링   → lol 관련 기능 클래스들을 domain.* 패키지 하위로 이동 - Spring 네이밍 규칙에 맞춰 패키지명을 단수형으로 변경   (ex. users → user, matches → match 등) - 사용되지 않는 클래스 및 불필요한 코드 제거
